🔨 Refactor:지현님 피드백 적용
