--- src/api/api.js
@@ -1,7 +1,0 @@
-import axios from 'axios';
-import { BACKEND_ENDPOINT } from '@/constants/endpoint';
-
-export const baseAxios = axios.create({
-  baseURL: BACKEND_ENDPOINT,
-  timeout: 20000,
-});
