--- src/api/index.js
@@ -1,6 +1,0 @@
-import axios from 'axios';
-
-export const baseAxios = axios.create({
-  baseURL: BACKEND_ENDPOINT,
-  timeout: 20000,
-});
